feature: Setup basic screens for CMP
