+++ composeApp/src/androidMain/kotlin/com/matijasokol/notes/MainActivity.kt
@@ -3,19 +3,13 @@ package com.matijasokol.notes
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.matijasokol.notes.ui.theme.NotesTheme
-import org.koin.compose.KoinContext
 
 class MainActivity : ComponentActivity() {
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     setContent {
-      KoinContext {
-        NotesTheme {
-          AppContent()
-        }
-      }
+      AppContent()
     }
   }
 }
