+++ composeApp/src/commonMain/kotlin/com/matijasokol/notes/navigation/Destination.kt
@@ -1,0 +1,12 @@
+package com.matijasokol.notes.navigation
+
+import kotlinx.serialization.Serializable
+
+sealed interface Destination {
+
+  @Serializable
+  data object List : Destination
+
+  @Serializable
+  data class Details(val noteId: Int) : Destination
+}
